Rearranged files
Moved the example game to game-eg, so game is now the starting point for authors of new games.
Custom exits are now defined in exits.js
Game files are listed in FILES in setting.js, so the author can add as many as needed.
The game folder is defined in page.html, in just one place