failedmsg
Use this for when the player tries to do something that fails. Output has the "failed" CSS class, so potentially can be differentiated - though that is not the case here yet.

Also the CanWearRemove and canTakeDrop are replaced by canManipulate, as there seems no practical difference

Several cmd rules renamed to make them shorter and more meaningful.