Setting now in their own namespace
This is a big change that will break any existing games (if there are any!).

There is a new setiings.js file in lib that has all the default settings. The settings.js file in the game directory, which is edited by authors, now only needs to include entries that change the default. This is neater, better JavaScript and should work better with the editor. Also biult-in support for custom libraries.

I have also modified the HTML file, with much of the work now done in a function in lib/setting.js.

Note that if your game currently has a `setup` function, that has to be in settings to run. Any other change will throw up an error so should be mpre obvious.

The only settings that are now required are "title" and "author".