Command objects uses "special"; liquids handled
The objects array for commands either specified the scope, or is flagged as "special". A "special" can be set to "ignore", "text" or "fluid", - or you can add your own. This means the existing ignore:true and text:true no longer work.

Also, liquid handling is officially included, complete with tests, albeit in a dumbed down manner.