The game object has been removed from w
This means it has a lot less unnecessary stuff that came from createObject, and no chance of a name collision if the author has a "game" object.
Most functionality has been moved to the world object, so all it contains is global variables for the game world pls three functions relating to saving.
Also added global variables "player" and "currentLocation", as these will be very convenient (previously I was thinking better not to have too many globals).
game.room is no longer available, but game.player is, at least for now, as it is used so much.
verbosity has moved to settings, findUniqueName moved to util