Updates to naming objects
properName is now properNoun, as that is the correct term.
parserAltNames is now synonyms

The second parameter for setAlias is now a dictionary, so you can set other attributes at the same time. It also is now used in createObject.
Added a getPlural to lang to support it.

Unrelated, the custom save/load methods have been renamed , again for consistency.