Reordering of start up dialog
There were two functions that are kicked off by the document loading, so some ambiguity in which fires first. There is just one, but I am slightly worried the order is now different, and something may not be quite right. However, it seems to work both with and without a dialog, so seems right.