Character action tests are now testXyz rather than canXyz
Character action tests are now testXyz rather than canXyz for consistency
Uncomments cannot_ingest in lang
GetAgreementGo now gets the exit, not just the direction. This required some reordering in the command, so if there is no exit, this will get noted in preference.
testMove is sent the exit, not the command (which was a bug)