Modified respond function and NPC reactions
This release significantly modified the "respond" function, which is used by other things. If you use the "respond" function with a default - the third parameter to the function - this will no longer work. You need to add an "extraScript" value to the first parameter. ASK/TELL has been updated to handle this.
The NPC reaction system now uses the "respond" function, which means that the "action" attribute needs to be changed to "script".