Minor updates
Walk-throughs can have functions in them. This allows you to set up special conditions while testing.
Items can have "sidebarButtonVerb" attribute; if set, clicking in the sidebar uses this verb, rather than displaying a choice.
The setItemAttThen agenda item does the same as setItemAtt, but the next item in the agenda is run immediately.
Tests and scripts in the "respond" code are bound to the "char" (so `this` will refer to the NPC), and are all sent the params and the response itself as two paramters.