Tweaks to random.dice and show
random.dice can be given a second parameter, true, an it will calculate the average, rather than roll at random
The "show" text processor directive has some extra options; there is also a "showOrNot" directive that can be given a default for when it evaluates to null or undefined. There are unit tests too.
The outer div for custom side panes are given an ID.