Throw error for clones created before initialising
There are issues if you try to clone an object before the object initialisation step (eg display verbs repeating). To avoid this, cloning should be done in settings.setup (or later). To make this a requirement, Quest will now throw an error during initialisation if it comes across a clone.